+++ 3_implementation/src/net/zebra/ls/hmm/Test.java
@@ -23,12 +23,12 @@ public class Test {
 		Factory factory = new FactoryImpl();
 		
 		DefaultHMM discreteHMM = ((FactoryImpl.HMMWrapper)factory.createDiscreteHMM(
-				new float[][] {
+				new double[][] {
 					{0.50f, 0.25f, 0.25f},
 					{0.30f, 0.40f, 0.30f},
 					{0.25f, 0.25f, 0.50f}}, 
-				new float[] {0.33f, 0.33f, 0.33f}, 
-				new float[][] {
+				new double[] {0.33f, 0.33f, 0.33f}, 
+				new double[][] {
 					{0.60f, 0.20f, 0.15f, 0.05f},
 					{0.25f, 0.25f, 0.25f, 0.25f},
 					{0.05f, 0.10f, 0.35f, 0.50f}})).getHMMImpl();
@@ -40,37 +40,37 @@ public class Test {
 				getHMMImpl();
 		
 		DefaultHMM normalHMM = ((FactoryImpl.HMMWrapper)factory.createNormalHMM(
-				new float[][] {
+				new double[][] {
 					{0.50f, 0.25f, 0.25f},
 					{0.30f, 0.40f, 0.30f},
 					{0.25f, 0.25f, 0.50f}}, 
-				new float[] {0.33f, 0.33f, 0.33f}, 
-				new float[] {0.87f, 0.14f, 0.39f}, 
-				new float[] {0.9f, 0.9f, 0.9f},
+				new double[] {0.33f, 0.33f, 0.33f}, 
+				new double[] {0.87f, 0.14f, 0.39f}, 
+				new double[] {0.9f, 0.9f, 0.9f},
 				Util.PROB_EPSILON)).getHMMImpl(); 
 		normalHMM.setStateNames(Arrays.asList("sunny", "cloudy", "rainy"));
 		
 		DefaultHMM exponentialHMM = ((FactoryImpl.HMMWrapper)factory.createExponentialHMM(
-				new float[][] {
+				new double[][] {
 					{0.50f, 0.25f, 0.25f},
 					{0.30f, 0.40f, 0.30f},
 					{0.25f, 0.25f, 0.50f}}, 
-				new float[] {0.33f, 0.33f, 0.33f}, 
-				new float[] {1.0f/0.87f, 1.0f/0.14f, 1.0f/0.39f},
+				new double[] {0.33f, 0.33f, 0.33f}, 
+				new double[] {1.0f/0.87f, 1.0f/0.14f, 1.0f/0.39f},
 				Util.PROB_EPSILON)).getHMMImpl(); 
 		exponentialHMM.setStateNames(Arrays.asList("sunny", "cloudy", "rainy"));
 		
 		DefaultHMM normalMixtureHMM = ((FactoryImpl.HMMWrapper)factory.createNormalMixtureHMM(
-				new float[][] {
+				new double[][] {
 					{0.50f, 0.25f, 0.25f},
 					{0.30f, 0.40f, 0.30f},
 					{0.25f, 0.25f, 0.50f}}, 
-				new float[] {0.33f, 0.33f, 0.33f}, 
-				new float[][] {
+				new double[] {0.33f, 0.33f, 0.33f}, 
+				new double[][] {
 					{0.87f, 0.15f}, {0.39f, 0.89f}, {0.14f, 0.37f}},
-				new float[][] {
+				new double[][] {
 					{1f, 1f}, {1f, 1f}, {1f, 1f}},
-				new float[][] {
+				new double[][] {
 					{0.6f, 0.4f}, {0.5f, 0.5f}, {0.4f, 0.6f}},
 				Util.PROB_EPSILON)).getHMMImpl();
 		normalMixtureHMM.setStateNames(Arrays.asList("sunny", "cloudy", "rainy"));
